activation Energy is now an UPPER limit & deep copying of parent coordinates
The founders now have energy for the founder Epitope that is below a
threshold. The founders have very low affinity.
When mutate is called, it used to copy the address of the parent and
then modify it so that all B cells had the same receptor. Now, mutate
copies the coordinates of the parent. It’s a deep copy.